refactor: simplify individual data fetching by introducing custom hooks
